+++ test/lib/jsdocMd.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const { strictEqual } = require('assert');
 const fs = require('fs');
-const { join } = require('path');
+const { join, resolve } = require('path');
 const { disposableDirectory } = require('disposable-directory');
+const snapshot = require('snapshot-assertion');
 const jsdocMd = require('../../lib/jsdocMd');
 
 module.exports = (tests) => {
@@ -110,106 +110,9 @@ Replace.
         targetHeading: 'Target',
       });
 
-      strictEqual(
+      await snapshot(
         await fs.promises.readFile(markdownPath, 'utf8'),
-        `# Preserve
-
-## Target
-
-### Table of contents
-
-- [class B](#class-b)
-  - [B static method d](#b-static-method-d)
-  - [B static property b](#b-static-property-b)
-  - [B instance method e](#b-instance-method-e)
-  - [B instance property c](#b-instance-property-c)
-  - [B inner typedef A](#b-inner-typedef-a)
-- [function c](#function-c)
-- [constant A](#constant-a)
-
-### class B
-
-Description, here is a **bold** word.
-
-| Parameter | Type     | Description                           |
-| :-------- | :------- | :------------------------------------ |
-| \`a\`       | boolean? | Description, here is a **bold** word. |
-
-#### Examples
-
-_Construct a new instance, here is a **bold** word._
-
-> \`\`\`js
-> const b = new B()
-> \`\`\`
-
-_Construct a new instance with options._
-
-> \`\`\`js
-> const b = new B(true)
-> \`\`\`
-
-#### B static method d
-
-Description.
-
-| Parameter | Type                      | Description  |
-| :-------- | :------------------------ | :----------- |
-| \`a\`       | [B~A](#b-inner-typedef-a) | Description. |
-| \`b\`       | boolean? = \`true\`         | Description. |
-
-#### B static property b
-
-Description.
-
-**Type:** string
-
-#### B instance method e
-
-Description.
-
-#### B instance property c
-
-Description.
-
-**Type:** string
-
-#### B inner typedef A
-
-Description.
-
-**Type:** Object
-
-| Property | Type    | Description                           |
-| :------- | :------ | :------------------------------------ |
-| \`a\`      | string  | Description, here is a **bold** word. |
-| \`b\`      | boolean | Description.                          |
-
-* * *
-
-### function c
-
-Description.
-
-| Parameter | Type   | Description  |
-| :-------- | :----- | :----------- |
-| \`a\`       | string | Description. |
-
-#### See
-
-- [\`B\`](#class-b).
-- [jsdoc-md on npm](https://npm.im/jsdoc-md).
-
-* * *
-
-### constant A
-
-Description.
-
-**Type:** boolean
-
-## Preserve
-`
+        resolve(__dirname, '../snapshots', 'jsdocMd.md')
       );
     });
   });