+++ changelog.md
@@ -7,6 +7,10 @@
 - Updated Node.js support to `^10.13.0 || ^12.0.0 || >= 13.7.0`.
 - Updated dev dependencies, some of which require newer Node.js versions than previously supported.
 
+### Patch
+
+- Converted a `jsdocMd` test assertion to a snapshot assertion.
+
 ## 6.0.0
 
 ### Major